requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.11.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
